hh_ru_parser_bot/requirements.txt
requests~=2.32.3
bs4~=0.0.2
beautifulsoup4~=4.12.3
aiogram~=3.8.0
python-dotenv~=1.0.1
SQLAlchemy~=2.0.31
asyncpg~=0.29.0
fake-useragent~=1.5.1
aiosqlite~=0.20.0

